server sending notifications to connected/disconnected users when they are added or removed from event

+++ app/socket-events/delete-event.js
@@ -5,23 +5,29 @@ module.exports = function(io, query) {
 				callback(false);
 				return;
 			}
-			query('delete from "event_user" where event_id = $1::int', [data.eventID], 
+			query('delete from "event_user" where event_id = $1::int returning user_id', [data.eventID], 
 				function(err, rows, result) {
 					if (err) {
 						callback(false);
 						return;
 					}
-					query('delete from "event" where event_id = $1::int', [data.eventID], 
-					function(err, rows, result) {
-						if (err) {
-							callback(false);
-							//return;
-						}
-						else {
-							callback(true);
-							//return;
-						}
-					});
+					var deletedUsers = rows;
+					query('delete from "notification" where event_id = $1::int', [data.eventID],
+						function(err, rows, result) {
+							if (err) {
+								callback(false);
+								return;
+							}
+							query('delete from "event" where event_id = $1::int', [data.eventID], 
+								function(err, rows, result) {
+									if (err) {
+										callback(false);
+									}
+									else {
+										callback(true, deletedUsers);
+									}
+								});
+						});
 			});
 		}
 	};
